Refresh token if needed
